refactor: 경험분해 feature에서 template 제거 => 리렌더링 이슈(언마운트, 마운트 됨)
